Create Probability when no of coins are thrown.
Here, we first set the no of times a coin is thrown and then accordingly get the no of outcomes using the no_of_turns function. therefore, we input the data in an array, and then, therefore, evaluate the no of heads and tails in the system, and then find the probability of both cases.